Refactor
- Add CSS folder
- Add components folder
- Clear all basic Vite React components (App.jsx)
